+++ audio_plugins/_SPARTA_matrixConv_/src/PluginEditor.h
@@ -96,11 +96,11 @@ private:
             durationInSeconds = (float)reader->lengthInSamples / (float)reader->sampleRate;
 
             if (reader->numChannels <= 1024 /* maximum number of channels for WAV files */) {
-                fileBuffer.setSize (reader->numChannels, (int) reader->lengthInSamples);
+                fileBuffer.setSize ((int)reader->numChannels, (int) reader->lengthInSamples);
                 reader->read (&fileBuffer, 0, (int) reader->lengthInSamples, 0, true, true);
             }
             const float** H = fileBuffer.getArrayOfReadPointers();
-            matrixconv_setFilters(hMC, H, fileBuffer.getNumChannels(), fileBuffer.getNumSamples(), reader->sampleRate);
+            matrixconv_setFilters(hMC, H, fileBuffer.getNumChannels(), fileBuffer.getNumSamples(), (int)reader->sampleRate);
         }
     }
